The ball rolls (v224)
Campbell: "it teleports to a different location. can we make it so that the ball
interpolates/moves between destinations to make it look like its rolling?"

Given real velocity and friction rather than lerping between waypoints, so the
roll falls out of the physics: a bat imparts speed AWAY from her (so the ball
leads the chase instead of doubling back through her), it coasts ~3.5 cells over
~4s, bounces off the room's walls, and she runs it down as it slows. Same
float x/y + vx/vy + prev_ shape as Critter/Firefly, so the renderer sub-tick
lerps it exactly like those.

Rolling read comes free: ball_roll accumulates with DISTANCE TRAVELLED, and the
highlight orbits by it — so the glint turns at the speed the ball is actually
moving and stops dead when it stops. No sprite needed.

She can only pounce on a ball that's slowed (BALL_REST_SPEED) and not twice in a
rally (BALL_BAT_COOLDOWN_TICKS), so one arrival is one bat rather than her
batting a ball still flying past her.

Also fixed, and it was masking the room: the Nest canvas took width/height as
JSX PROPS, so every re-render (the stats poll, 4x a second) reset them and
silently undid the imperative resize to the room's dimensions — the crop was
right but drawn into a stale-sized canvas. Canvas size is React state now.

Test: 28 checks. Biggest single-tick move is 6px (a roll) where a teleport was
48px, and the ball travels ~194px over a chase. v224 serial-clean 120s on
hardware; browser shows bond 0->13% and loneliness 80->1 across one game.

+++ host/host_web.cpp
@@ -624,8 +624,8 @@ EMSCRIPTEN_KEEPALIVE int host_room_h() { Chamber& ch = chamber(); return (ch.roo
 // Ball position in display pixels, -1 when nothing's in play. Exists so a test can
 // assert the ball stays inside the Nest's window — the first version threw it out
 // of view every time and looked, to a keeper, like the button did nothing.
-EMSCRIPTEN_KEEPALIVE int host_pr_ball_x() { Chamber& ch = chamber(); return ch.has_ball() ? ch.ball_x * Cfg::CELL_SIZE : -1; }
-EMSCRIPTEN_KEEPALIVE int host_pr_ball_y() { Chamber& ch = chamber(); return ch.has_ball() ? ch.ball_y * Cfg::CELL_SIZE : -1; }
+EMSCRIPTEN_KEEPALIVE int host_pr_ball_x() { Chamber& ch = chamber(); return ch.has_ball() ? (int)(ch.ball_x * Cfg::CELL_SIZE) : -1; }
+EMSCRIPTEN_KEEPALIVE int host_pr_ball_y() { Chamber& ch = chamber(); return ch.has_ball() ? (int)(ch.ball_y * Cfg::CELL_SIZE) : -1; }
 EMSCRIPTEN_KEEPALIVE int host_pr_sleeping() { Chamber& ch = chamber(); return ch.conker_count ? (ch.conkers[0].sleeping?1:0) : 0; }
 // Live needs (0..100), all active per NEEDS_ACTIVE_MASK — the app shows them as bars.
 EMSCRIPTEN_KEEPALIVE int host_pr_boredom() { Chamber& ch = chamber(); return ch.conker_count ? (int)(ch.conkers[0].needs[NEED_BOREDOM]*100.0f) : 0; }
